+++ modules/sgct/include/inviwo/sgct/datastructures/sgctcamera.h
@@ -42,12 +42,13 @@ namespace inviwo {
  */
 class IVW_MODULE_SGCT_API SGCTCamera final : public Camera {
 public:
-    explicit SGCTCamera(vec3 lookFrom = cameradefaults::lookFrom,
-                        vec3 lookTo = cameradefaults::lookTo, vec3 lookUp = cameradefaults::lookUp,
-                        float nearPlane = cameradefaults::nearPlane,
-                        float farPlane = cameradefaults::farPlane,
-                        float aspectRatio = cameradefaults::aspectRatio,
-                        float fieldOfView = cameradefaults::fieldOfView);
+    explicit SGCTCamera(dvec3 lookFrom = cameradefaults::lookFrom,
+                        dvec3 lookTo = cameradefaults::lookTo,
+                        dvec3 lookUp = cameradefaults::lookUp,
+                        double nearPlane = cameradefaults::nearPlane,
+                        double farPlane = cameradefaults::farPlane,
+                        double aspectRatio = cameradefaults::aspectRatio,
+                        double fieldOfView = cameradefaults::fieldOfView);
 
     SGCTCamera(const SGCTCamera& other) = default;
     SGCTCamera& operator=(const SGCTCamera& other) = default;
@@ -62,8 +63,8 @@ public:
     virtual void updateFrom(const Camera& source) override;
     virtual void configureProperties(CameraProperty& cameraProperty, bool attach) override;
 
-    float getFovy() const;
-    void setFovy(float val);
+    double getFovy() const;
+    void setFovy(double val);
     void setExternal(const sgct::RenderData& renderData);
     virtual void zoom(const ZoomOptions& opts) override;
 
@@ -72,16 +73,16 @@ public:
 
 protected:
     virtual bool equal(const Camera& other) const override;
-    virtual mat4 calculateProjectionMatrix() const override;
-    virtual mat4 calculateViewMatrix() const override;
+    virtual dmat4 calculateProjectionMatrix() const override;
+    virtual dmat4 calculateViewMatrix() const override;
 
-    float fovy_;
+    double fovy_;
 
-    std::optional<mat4> extProj_{};
-    mat4 extView_{1.0f};
-    mat4 extModel_{1.0f};
+    std::optional<dmat4> extProj_{};
+    dmat4 extView_{1.0};
+    dmat4 extModel_{1.0};
 };
 
-inline float SGCTCamera::getFovy() const { return fovy_; }
+inline double SGCTCamera::getFovy() const { return fovy_; }
 
 }  // namespace inviwo
